+++ src/Kiota.Builder/Logging/FileLogLoggerProvider.cs
@@ -4,16 +4,11 @@ using Microsoft.Extensions.Logging;
 
 namespace Kiota.Builder.Logging;
 
-public sealed class FileLogLoggerProvider : ILoggerProvider
+public sealed class FileLogLoggerProvider(string logFileDirectoryAbsolutePath, LogLevel logLevel) : ILoggerProvider
 {
-    private readonly LogLevel _logLevel;
-    private readonly string _logFileDirectoryAbsolutePath;
-    public FileLogLoggerProvider(string logFileDirectoryAbsolutePath, LogLevel logLevel)
-    {
-        _logLevel = logLevel;
-        _logFileDirectoryAbsolutePath = logFileDirectoryAbsolutePath;
-    }
-    private readonly List<FileLogLogger> _loggers = new();
+    private readonly LogLevel _logLevel = logLevel;
+    private readonly string _logFileDirectoryAbsolutePath = logFileDirectoryAbsolutePath;
+    private readonly List<FileLogLogger> _loggers = [];
     public ILogger CreateLogger(string categoryName)
     {
         var instance = new FileLogLogger(_logFileDirectoryAbsolutePath, _logLevel, categoryName);