+++ vscode/microsoft-kiota/src/commands/openApidescription/searchOrOpenApiDescriptionCommand.ts
@@ -1,11 +1,12 @@
 import TelemetryReporter from "@vscode/extension-telemetry";
 import * as vscode from "vscode";
 
-import { extensionId, treeViewId } from "../../constants";
+import { extensionId, SHOW_MESSAGE_AFTER_API_LOAD, treeViewId } from "../../constants";
 import { setDeepLinkParams } from "../../handlers/deepLinkParamsHandler";
 import { searchSteps } from "../../modules/steps/searchSteps";
 import { OpenApiTreeProvider } from "../../providers/openApiTreeProvider";
 import { getExtensionSettings } from "../../types/extensionSettings";
+import { updateTreeViewIcons } from "../../util";
 import { IntegrationParams, validateDeepLinkQueryParams } from "../../utilities/deep-linking";
 import { openTreeViewWithProgress } from "../../utilities/progress";
 import { Command } from "../Command";
@@ -13,13 +14,11 @@ import { searchDescription } from "./searchDescription";
 
 export class SearchOrOpenApiDescriptionCommand extends Command {
 
-  private _openApiTreeProvider: OpenApiTreeProvider;
-  private _context: vscode.ExtensionContext;
-
-  constructor(openApiTreeProvider: OpenApiTreeProvider, context: vscode.ExtensionContext) {
+  constructor(
+    private openApiTreeProvider: OpenApiTreeProvider,
+    private context: vscode.ExtensionContext
+  ) {
     super();
-    this._openApiTreeProvider = openApiTreeProvider;
-    this._context = context;
   }
 
   public getName(): string {
@@ -31,7 +30,7 @@ export class SearchOrOpenApiDescriptionCommand extends Command {
     if (Object.keys(searchParams).length > 0) {
       let [params, errorsArray] = validateDeepLinkQueryParams(searchParams);
       setDeepLinkParams(params);
-      const reporter = new TelemetryReporter(this._context.extension.packageJSON.telemetryInstrumentationKey);
+      const reporter = new TelemetryReporter(this.context.extension.packageJSON.telemetryInstrumentationKey);
       reporter.sendTelemetryEvent("DeepLinked searchOrOpenApiDescription", {
         "searchParameters": JSON.stringify(searchParams),
         "validationErrors": errorsArray.join(", ")
@@ -40,7 +39,7 @@ export class SearchOrOpenApiDescriptionCommand extends Command {
 
     // proceed to enable loading of openapi description
     const yesAnswer = vscode.l10n.t("Yes, override it");
-    if (this._openApiTreeProvider.hasChanges()) {
+    if (this.openApiTreeProvider.hasChanges()) {
       const response = await vscode.window.showWarningMessage(
         vscode.l10n.t(
           "Before adding a new API description, consider that your changes and current selection will be lost."),
@@ -58,12 +57,32 @@ export class SearchOrOpenApiDescriptionCommand extends Command {
       title: vscode.l10n.t("Searching...")
     }, (progress, _) => {
       const settings = getExtensionSettings(extensionId);
-      return searchDescription(this._context, x, settings.clearCache);
+      return searchDescription(this.context, x, settings.clearCache);
     }));
 
     if (config.descriptionPath) {
-      await openTreeViewWithProgress(() => this._openApiTreeProvider.setDescriptionUrl(config.descriptionPath!));
+      await openTreeViewWithProgress(async () => {
+        await this.openApiTreeProvider.setDescriptionUrl(config.descriptionPath!);
+        await updateTreeViewIcons(treeViewId, true, false);
+      });
+
+      const generateAnswer = vscode.l10n.t("Generate");
+      const showGenerateMessage = this.context.globalState.get<boolean>(SHOW_MESSAGE_AFTER_API_LOAD, true);
+
+      if (showGenerateMessage) {
+        const doNotShowAgainOption = vscode.l10n.t("Do not show this again");
+        const response = await vscode.window.showInformationMessage(
+          vscode.l10n.t('Click on Generate after selecting the paths in the API Explorer'),
+          generateAnswer,
+          doNotShowAgainOption
+        );
+
+        if (response === generateAnswer) {
+          await vscode.commands.executeCommand(`${treeViewId}.generateClient`);
+        } else if (response === doNotShowAgainOption) {
+          await this.context.globalState.update(SHOW_MESSAGE_AFTER_API_LOAD, false);
+        }
+      }
     }
-    await vscode.window.showInformationMessage(vscode.l10n.t('You can now select the required endpoints from {0}', this._openApiTreeProvider.apiTitle!));
   }
 }