+++ vscode/microsoft-kiota/src/util.ts
@@ -171,3 +171,5 @@ export function isValidUrl(url: string): boolean {
     return false;
   }
 }
+
+