+++ vscode/microsoft-kiota/CHANGELOG.md
@@ -8,6 +8,13 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+## [1.22.100000001] - 2025-01-10
+
+### Added
+
+### Changed
+- Fixed a bug in the VS Code extension plugin generation [#5978](https://github.com/microsoft/kiota/issues/5978)
+
 ## [1.20.000000001] - 2024-11-07
 
 ## [1.19.100000001] - 2024-10-03